+++ src/main/java/com/swordglowsblue/artifice/api/ArtificeResourcePack.java
@@ -9,7 +9,7 @@ import com.swordglowsblue.artifice.api.builder.assets.ModelBuilder;
 import com.swordglowsblue.artifice.api.builder.assets.ParticleBuilder;
 import com.swordglowsblue.artifice.api.builder.assets.TranslationBuilder;
 import com.swordglowsblue.artifice.api.builder.data.AdvancementBuilder;
-import com.swordglowsblue.artifice.api.builder.data.biome.BiomeBuilder;
+import com.swordglowsblue.artifice.api.builder.data.worldgen.biome.BiomeBuilder;
 import com.swordglowsblue.artifice.api.builder.data.dimension.DimensionBuilder;
 import com.swordglowsblue.artifice.api.builder.data.dimension.DimensionTypeBuilder;
 import com.swordglowsblue.artifice.api.builder.data.LootTableBuilder;
@@ -280,7 +280,7 @@ public interface ArtificeResourcePack extends ResourcePack, ServerResourcePackPr
          * Add a Biome with the given ID.
          *
          * @param id The ID of the biome, which will be converted into the correct path.
-         * @param f A callback which will be passed an {@link com.swordglowsblue.artifice.api.builder.data.biome.BiomeBuilder} to create the biome .
+         * @param f A callback which will be passed an {@link com.swordglowsblue.artifice.api.builder.data.worldgen.biome.BiomeBuilder} to create the biome .
          */
         void addBiome(Identifier id, Processor<BiomeBuilder> f);
