+++ src/main/java/com/swordglowsblue/artifice/api/builder/data/worldgen/biome/BiomeBuilder.java
@@ -1,4 +1,4 @@
-package com.swordglowsblue.artifice.api.builder.data.biome;
+package com.swordglowsblue.artifice.api.builder.data.worldgen.biome;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
